Split out sub-classes.
